% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11DataTypes-package.R
\docType{package}
\name{cpp11DataTypes-package}
\alias{cpp11DataTypes}
\alias{cpp11DataTypes-package}
\title{cpp11DataTypes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
